Add QR code in Address Details modal
